ignore http commands code coverage
